pyproject.toml
[project]
name = "calibration-lifelog"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ace-tools>=0.0",
    "bunnet>=1.3.0",
    "fastapi[standard]>=0.115.12",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.3",
    "pandas>=2.2.3",
    "pendulum>=3.1.0",
    "plotly>=6.1.1",
    "pymongo>=4.13.0",
    "python-dotenv>=1.0.0",
    "requests>=2.32.3",
]
